Simplifing tests
moving test to its own file
