+++ Assets/Mirror/Tests/Editor/SyncListStrucTest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85fb5738f517c4c4a93c3af85261e67e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
